Updated to compile w/o warnings on new PS2SDK revision, modified GUI to support loading of embedded resources for the internal theme, added menu hints, added an option for changing the select button.

+++ modules/hdd/ps2fs/pfs_fio.c
@@ -690,7 +690,7 @@ s64 _seek(pfs_file_slot_t *fileSlot, s64 offset, int whence, int mode)
 	return newPos;
 }
 
-int pfsLseek(iop_file_t *f, unsigned long pos, int whence)
+int pfsLseek(iop_file_t *f, int pos, int whence)
 {
 	pfs_file_slot_t *fileSlot = (pfs_file_slot_t*)f->privdata;
 	int result = checkFileSlot(fileSlot);
